backend/requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Restless-NG==2.2.9
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.32
marshmallow==3.14.1
flask-marshmallow==0.14.0
requests==2.27.1
python-dotenv==0.19.2
marshmallow-sqlalchemy==0.27.0
psycopg2==2.9.3
uWSGI==2.0.20
